windows changes.

+++ jde/markets/types/Tick.h
@@ -58,7 +58,7 @@ namespace Jde::Markets
 		bool SetString( ETickType type, str value )noexcept;
 		bool SetInt( ETickType type, _int value )noexcept;
 		bool SetPrice( ETickType type, double value/*, const ::TickAttrib& attribs*/ )noexcept;
-		void SetPrices( double bidSize, double bid, double askSize, double ask )noexcept;
+		void SetPrices( long long bidSize, double bid, long long askSize, double ask )noexcept;
 		bool SetDouble( ETickType type, double value )noexcept;
 		bool SetOptionComputation( ETickType type, OptionComputation&& v )noexcept;
 		bool FieldEqual( const Tick& other, ETickType tick )const noexcept;
@@ -78,7 +78,7 @@ namespace Jde::Markets
 		double Ask;
 		long long AskSize;
 		double LastPrice;
-		double LastSize;
+		long long LastSize;
 		double High;
 		double Low;
 		uint Volume;
